modify DTO
