--- app/src/main/java/myapplication/second/workinghourmanagement/dto/manageStaff/FindStaff.kt
@@ -1,7 +1,0 @@
-package myapplication.second.workinghourmanagement.dto.manageStaff
-
-data class FindStaff(
-    val staffId: Int,
-    val phone: String,
-    val name: String,
-)
