+++ app/src/main/java/myapplication/second/workinghourmanagement/dto/manageStaff/ResponseGetStaffInfo.kt
@@ -8,3 +8,21 @@ data class ResponseGetStaffInfo(
     val statusCode: Int,
     val timestamp: String
 )
+
+data class StaffInfo(
+    val employmentId: Int,
+    val staffId: Int,
+    val name: String,
+    val phone: String,
+    val profileImage: String?,
+    val rank: String?,
+    val hourlyWage: Int?,
+    val readWorkTimesWithStaffList: List<WorkTime>
+)
+
+data class WorkTime(
+    val workTimeId: Int,
+    val dayOfWeek: String,
+    val startTime: String,
+    val endTime: String
+)
