+++ app/src/main/java/myapplication/second/workinghourmanagement/dto/manageStaff/StaffList.kt
@@ -4,7 +4,17 @@ data class StaffList(
     val path: String,
     val method: String,
     val message: String,
-    val data: List<Staff>,
+    val data: Map<String, List<Staff>>,
     val statusCode: Int,
     val timestamp: String
+)
+
+data class Staff(
+    val employmentId: Int,
+    val staffName: String,
+    val profileImage: String?,
+    val day: String?,
+    val startTime: String?,
+    val endTime: String?,
+    val attendanceStatus: String
 )
