--- app/src/main/java/myapplication/second/workinghourmanagement/dto/manageStaff/WorkTimeRequest.kt
@@ -1,9 +1,0 @@
-package myapplication.second.workinghourmanagement.dto.manageStaff
-
-data class WorkTimeRequest(
-    val dayOfWeekList: List<String>,
-    val startDate: String,
-    val endDate: String,
-    val startTime: String,
-    val endTime: String
-)
